+++ src/daemon/interactive_command_executor.h
@@ -42,7 +42,7 @@ public:
 
   bool print_transaction_pool_short() override;
 
-  bool start_mining(cryptonote::account_public_address address, size_t num_threads) override;
+  bool start_mining(cryptonote::account_public_address address, uint64_t num_threads) override;
 
   bool stop_mining() override;
 };
